refactor: application -> facade
