+++ src/main/kotlin/gomushin/backend/member/dto/request/OnboardingRequest.kt
@@ -1,4 +1,4 @@
-package gomushin.backend.member.presentation.dto.request
+package gomushin.backend.member.dto.request
 
 import io.swagger.v3.oas.annotations.media.Schema
 import java.time.LocalDate
