+++ src/main/kotlin/gomushin/backend/member/presentation/MemberInfoController.kt
@@ -2,8 +2,8 @@ package gomushin.backend.member.presentation
 
 import gomushin.backend.core.CustomUserDetails
 import gomushin.backend.core.common.web.response.ApiResponse
-import gomushin.backend.member.application.MemberInfoFacade
-import gomushin.backend.member.presentation.dto.response.MyInfoResponse
+import gomushin.backend.member.facade.MemberInfoFacade
+import gomushin.backend.member.dto.response.MyInfoResponse
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.tags.Tag
 import org.springframework.security.core.annotation.AuthenticationPrincipal
